+++ CountlyServerConfig.m
@@ -6,12 +6,71 @@
 
 #import "CountlyCommon.h"
 
-@interface CountlyServerConfig ()
+@interface CountlyServerConfig () {
+    NSTimer *_requestTimer;
+}
 @property (nonatomic) BOOL trackingEnabled;
 @property (nonatomic) BOOL networkingEnabled;
+@property (nonatomic) BOOL crashReportingEnabled;
+@property (nonatomic) BOOL loggingEnabled;
+@property (nonatomic) BOOL customEventTrackingEnabled;
+@property (nonatomic) BOOL viewTrackingEnabled;
+@property (nonatomic) BOOL sessionTrackingEnabled;
+@property (nonatomic) BOOL enterContentZone;
+@property (nonatomic) BOOL consentRequired;
+@property (nonatomic) BOOL locationTracking;
+@property (nonatomic) BOOL refreshContentZone;
+
+@property (nonatomic) NSInteger limitKeyLength;
+@property (nonatomic) NSInteger limitValueSize;
+@property (nonatomic) NSInteger limitSegValues;
+@property (nonatomic) NSInteger limitBreadcrumb;
+@property (nonatomic) NSInteger limitTraceLine;
+@property (nonatomic) NSInteger limitTraceLength;
+@property (nonatomic) NSInteger sessionInterval;
+@property (nonatomic) NSInteger eventQueueSize;
+@property (nonatomic) NSInteger requestQueueSize;
+@property (nonatomic) NSInteger contentZoneInterval;
+@property (nonatomic) NSInteger dropOldRequestTime;
+@property (nonatomic) NSInteger serverConfigUpdateInterval;
+@property (nonatomic) NSInteger currentServerConfigUpdateInterval;
+
+@property (nonatomic) NSInteger version;
+@property (nonatomic) long long timestamp;
+@property (nonatomic) long long lastFetchTimestamp;
+
 @end
 
 NSString* const kCountlySCKeySC = @"sc";
+NSString* const kTracking = @"tracking";
+NSString* const kNetworking = @"networking";
+
+// request keys
+NSString* const kRTimestamp = @"t";
+NSString* const kRVersion = @"v";
+NSString* const kRConfig = @"c";
+NSString* const kRReqQueueSize = @"rqs";
+NSString* const kREventQueueSize = @"eqs";
+NSString* const kRLogging = @"log";
+NSString* const kRSessionUpdateInterval = @"sui";
+NSString* const kRSessionTracking = @"st";
+NSString* const kRViewTracking = @"vt";
+NSString* const kRLocationTracking = @"lt";
+NSString* const kRRefreshContentZone = @"rcz";
+
+NSString* const kRLimitKeyLength = @"lkl";
+NSString* const kRLimitValueSize = @"lvs";
+NSString* const kRLimitSegValues = @"lsv";
+NSString* const kRLimitBreadcrumb = @"lbc";
+NSString* const kRLimitTraceLine = @"ltlpt";
+NSString* const kRLimitTraceLength = @"ltl";
+NSString* const kRCustomEventTracking = @"cet";
+NSString* const kREnterContentZone = @"ecz";
+NSString* const kRContentZoneInterval = @"czi";
+NSString* const kRConsentRequired = @"cr";
+NSString* const kRDropOldRequestTime = @"dort";
+NSString* const kRCrashReporting = @"crt";
+NSString* const kRServerConfigUpdateInterval = @"scui";
 
 @implementation CountlyServerConfig
 
@@ -28,61 +87,230 @@ NSString* const kCountlySCKeySC = @"sc";
 
 - (instancetype)init
 {
-    if (self = [super init])
-    {
-        self.trackingEnabled = YES;
-        self.networkingEnabled = YES;
-        NSDictionary* serverConfigObject = [CountlyPersistency.sharedInstance retrieveServerConfig];
-        if (serverConfigObject) {
-            [self populateServerConfig:serverConfigObject];
-        }
+    self = [super init];
+    if (self) {
+        // Set default values
+        _trackingEnabled = YES;
+        _networkingEnabled = YES;
+        _crashReportingEnabled = YES;
+        _customEventTrackingEnabled = YES;
+        _enterContentZone = NO;
+        _locationTracking= YES;
+        _viewTrackingEnabled = YES;
+        _sessionTrackingEnabled = YES;
+        _loggingEnabled = NO;
+        _refreshContentZone = YES;
+        
+        _timestamp = 0;
+        _version = 0;
+        _currentServerConfigUpdateInterval = 4;
+        _requestTimer = nil;
     }
-    
     return self;
 }
 
-- (BOOL)trackingEnabled
+- (void)retrieveServerConfigFromStorage:(NSString*) providedServerConfiguration
 {
-    if (!CountlyCommon.sharedInstance.enableServerConfiguration)
-        return YES;
-    
-    return _trackingEnabled;
-}
-
-- (BOOL)networkingEnabled
-{
-    if (!CountlyCommon.sharedInstance.enableServerConfiguration)
-        return YES;
-    
-    return _networkingEnabled;
-}
-
-- (void)populateServerConfig:(NSDictionary *)dictionary
-{
-    if (dictionary[@"tracking"])
-    {
-        self.trackingEnabled = [dictionary[@"tracking"] boolValue];
-    }
-    if (dictionary[@"networking"])
-    {
-        self.networkingEnabled = [dictionary[@"networking"] boolValue];
+    NSError* error = nil;
+    NSDictionary* serverConfigObject;
+    if(providedServerConfiguration) {
+        serverConfigObject = [NSJSONSerialization JSONObjectWithData:[providedServerConfiguration cly_dataUTF8] options:0 error:&error];
     }
     
-    CLY_LOG_D(@"tracking : %@", self.trackingEnabled ? @"YES" : @"NO");
-    CLY_LOG_D(@"networking : %@", self.networkingEnabled ? @"YES" : @"NO");
+    if(error){
+        serverConfigObject = [CountlyPersistency.sharedInstance retrieveServerConfig];
+    }
+    
+    if (serverConfigObject) {
+        [self populateServerConfig:serverConfigObject];
+    }
 }
 
-- (void)fetchServerConfig
+
+- (void)setBoolProperty:(BOOL *)property fromDictionary:(NSDictionary *)dictionary key:(NSString *)key logString:(NSMutableString *)logString
+{
+    NSNumber *value = dictionary[key];
+    if (value) {
+        *property = value.boolValue;
+        [logString appendFormat:@"%@: %@, ", key, *property ? @"YES" : @"NO"];
+    }
+}
+
+- (void)setIntegerProperty:(NSInteger *)property fromDictionary:(NSDictionary *)dictionary key:(NSString *)key logString:(NSMutableString *)logString
+{
+    NSNumber *value = dictionary[key];
+    if (value) {
+        *property = value.integerValue;
+        [logString appendFormat:@"%@: %ld, ", key, (long)*property];
+    }
+}
+
+- (void)populateServerConfig:(NSDictionary *)serverConfig
+{
+    if(!serverConfig[kRConfig]) {
+        CLY_LOG_D(@"%s, config key is missing in the server configuration omitting", __FUNCTION__);
+        return;
+    }
+    
+    NSDictionary* dictionary = serverConfig[kRConfig];
+    
+    if(!serverConfig[kRVersion] || !serverConfig[kRTimestamp]) {
+        CLY_LOG_D(@"%s, version or timestamp is missing in the server configuration omitting", __FUNCTION__);
+        return;
+    }
+    
+    _version = [serverConfig[kRVersion] integerValue];
+    _timestamp = [serverConfig[kRTimestamp] longLongValue];
+    
+    NSMutableString *logString = [NSMutableString stringWithString:@"Server Config: "];
+    
+    [self setBoolProperty:&_trackingEnabled fromDictionary:dictionary key:kTracking logString:logString];
+    [self setBoolProperty:&_networkingEnabled fromDictionary:dictionary key:kNetworking logString:logString];
+    [self setIntegerProperty:&_sessionInterval fromDictionary:dictionary key:kRSessionUpdateInterval logString:logString];
+    [self setIntegerProperty:&_requestQueueSize fromDictionary:dictionary key:kRReqQueueSize logString:logString];
+    [self setIntegerProperty:&_eventQueueSize fromDictionary:dictionary key:kREventQueueSize logString:logString];
+    [self setBoolProperty:&_crashReportingEnabled fromDictionary:dictionary key:kRCrashReporting logString:logString];
+    [self setBoolProperty:&_sessionTrackingEnabled fromDictionary:dictionary key:kRSessionTracking logString:logString];
+    [self setBoolProperty:&_loggingEnabled fromDictionary:dictionary key:kRLogging logString:logString];
+    [self setIntegerProperty:&_limitKeyLength fromDictionary:dictionary key:kRLimitKeyLength logString:logString];
+    [self setIntegerProperty:&_limitValueSize fromDictionary:dictionary key:kRLimitValueSize logString:logString];
+    [self setIntegerProperty:&_limitSegValues fromDictionary:dictionary key:kRLimitSegValues logString:logString];
+    [self setIntegerProperty:&_limitBreadcrumb fromDictionary:dictionary key:kRLimitBreadcrumb logString:logString];
+    [self setIntegerProperty:&_limitTraceLine fromDictionary:dictionary key:kRLimitTraceLine logString:logString];
+    [self setIntegerProperty:&_limitTraceLength fromDictionary:dictionary key:kRLimitTraceLength logString:logString];
+    [self setBoolProperty:&_customEventTrackingEnabled fromDictionary:dictionary key:kRCustomEventTracking logString:logString];
+    [self setBoolProperty:&_viewTrackingEnabled fromDictionary:dictionary key:kRViewTracking logString:logString];
+    [self setBoolProperty:&_enterContentZone fromDictionary:dictionary key:kREnterContentZone logString:logString];
+    [self setIntegerProperty:&_contentZoneInterval fromDictionary:dictionary key:kRContentZoneInterval logString:logString];
+    [self setBoolProperty:&_consentRequired fromDictionary:dictionary key:kRConsentRequired logString:logString];
+    [self setIntegerProperty:&_dropOldRequestTime fromDictionary:dictionary key:kRDropOldRequestTime logString:logString];
+    [self setIntegerProperty:&_serverConfigUpdateInterval fromDictionary:dictionary key:kRServerConfigUpdateInterval logString:logString];
+    [self setBoolProperty:&_locationTracking fromDictionary:dictionary key:kRLocationTracking logString:logString];
+    [self setBoolProperty:&_refreshContentZone fromDictionary:dictionary key:kRRefreshContentZone logString:logString];
+
+    CLY_LOG_D(@"%s, version:[%li], timestamp:[%lli], %@", __FUNCTION__, _version, _timestamp, logString);
+}
+
+- (void)notifySdkConfigChange:(CountlyConfig *)config
+{
+    config.enableDebug = _loggingEnabled || config.enableDebug;
+    CountlyCommon.sharedInstance.enableDebug = config.enableDebug;
+    
+    // Limits could be moved to another function, but letting them stay here serves us a monopolized view of notify
+    if (config.maxKeyLength) {
+        [config.sdkInternalLimits setMaxKeyLength: config.maxKeyLength];
+    }
+    
+    if (config.maxValueLength) {
+        [config.sdkInternalLimits setMaxValueSize: config.maxValueLength];
+    }
+    
+    if (config.maxSegmentationValues) {
+        [config.sdkInternalLimits setMaxSegmentationValues: config.maxSegmentationValues];
+    }
+    
+    if (config.crashLogLimit) {
+        [config.sdkInternalLimits setMaxBreadcrumbCount: config.crashLogLimit];
+    }
+    
+    [config.sdkInternalLimits setMaxKeyLength: _limitKeyLength ?: config.sdkInternalLimits.getMaxKeyLength];
+    [config.sdkInternalLimits setMaxValueSize: _limitValueSize ?: config.sdkInternalLimits.getMaxValueSize];
+    [config.sdkInternalLimits setMaxSegmentationValues: _limitSegValues ?: config.sdkInternalLimits.getMaxSegmentationValues];
+    [config.sdkInternalLimits setMaxBreadcrumbCount: _limitBreadcrumb ?: config.sdkInternalLimits.getMaxBreadcrumbCount];
+    [config.sdkInternalLimits setMaxStackTraceLineLength: _limitTraceLength ?: config.sdkInternalLimits.getMaxStackTraceLineLength];
+    [config.sdkInternalLimits setMaxStackTraceLinesPerThread: _limitTraceLine ?: config.sdkInternalLimits.getMaxStackTraceLinesPerThread];
+    
+    CountlyCommon.sharedInstance.maxKeyLength = config.sdkInternalLimits.getMaxKeyLength;
+    CountlyCommon.sharedInstance.maxValueLength = config.sdkInternalLimits.getMaxValueSize;
+    CountlyCommon.sharedInstance.maxSegmentationValues = config.sdkInternalLimits.getMaxSegmentationValues;
+    
+    config.requiresConsent = _consentRequired ?: config.requiresConsent;
+    CountlyConsentManager.sharedInstance.requiresConsent = config.requiresConsent;
+    
+    config.eventSendThreshold = _eventQueueSize ?: config.eventSendThreshold;
+    config.requestDropAgeHours = _dropOldRequestTime ?: config.requestDropAgeHours;
+    config.storedRequestsLimit = _requestQueueSize ?: config.storedRequestsLimit;
+    CountlyPersistency.sharedInstance.eventSendThreshold = config.eventSendThreshold;
+    CountlyPersistency.sharedInstance.requestDropAgeHours = config.requestDropAgeHours;
+    CountlyPersistency.sharedInstance.storedRequestsLimit = MAX(1, config.storedRequestsLimit);
+    
+    config.updateSessionPeriod = _sessionInterval ?: config.updateSessionPeriod;
+    
+    [config.content setZoneTimerInterval: _contentZoneInterval ?: config.content.getZoneTimerInterval];
+    
+    CountlyContentBuilderInternal.sharedInstance.zoneTimerInterval = config.content.getZoneTimerInterval;
+    CountlyCrashReporter.sharedInstance.crashLogLimit = config.sdkInternalLimits.getMaxBreadcrumbCount;
+    
+    if(!_enterContentZone){
+        dispatch_async(dispatch_get_main_queue(), ^{
+            [CountlyContentBuilderInternal.sharedInstance exitContentZone];
+        });
+    } else {
+        dispatch_async(dispatch_get_main_queue(), ^{
+            [CountlyContentBuilderInternal.sharedInstance exitContentZone];
+            [CountlyContentBuilderInternal.sharedInstance enterContentZone:@[]];
+        });
+    }
+    
+    if(_serverConfigUpdateInterval && _serverConfigUpdateInterval != _currentServerConfigUpdateInterval && _requestTimer){
+        _currentServerConfigUpdateInterval = _serverConfigUpdateInterval;
+        [_requestTimer invalidate];
+        _requestTimer = nil;
+        _requestTimer = [NSTimer timerWithTimeInterval:_currentServerConfigUpdateInterval * 60 * 60
+                                                         target:self selector:@selector(fetchServerConfigTimer:) userInfo:config repeats:YES];
+        [NSRunLoop.mainRunLoop addTimer:_requestTimer forMode:NSRunLoopCommonModes];
+    }
+    
+    if(!_locationTracking && !CountlyLocationManager.sharedInstance.isLocationInfoDisabled){
+        [CountlyLocationManager.sharedInstance disableLocationInfo];
+    }
+}
+
+- (void)fetchServerConfigTimer:(NSTimer *)timer
+{
+    CountlyConfig *config = (CountlyConfig *)timer.userInfo; // Retrieve CountlyConfig from userInfo
+    if (config) {
+        [self fetchServerConfig:config];
+    }
+}
+
+- (void)fetchServerConfigIfTimeIsUp
+{
+    if(_lastFetchTimestamp) {
+        long long currentTime = NSDate.date.timeIntervalSince1970 * 1000;
+        long long timePassed = currentTime - _lastFetchTimestamp;
+        
+        if(timePassed > _currentServerConfigUpdateInterval * 60 * 60 * 1000){
+            [self fetchServerConfig:CountlyConfig.new];
+        }
+    }
+}
+
+- (void)fetchServerConfig:(CountlyConfig *)config
 {
     CLY_LOG_D(@"Fetching server configs...");
-    if (!CountlyCommon.sharedInstance.enableServerConfiguration)
-    {
-        CLY_LOG_D(@"'fetchServerConfig' enable server configuration during init time configuration.");
-        return;
-    }
-    
     if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
         return;
+    
+    _lastFetchTimestamp = NSDate.date.timeIntervalSince1970 * 1000;
+    
+    if(!_requestTimer){
+        _requestTimer = [NSTimer timerWithTimeInterval:_currentServerConfigUpdateInterval * 60 * 60
+                                                         target:self selector:@selector(fetchServerConfigTimer:) userInfo:config repeats:YES];
+        [NSRunLoop.mainRunLoop addTimer:_requestTimer forMode:NSRunLoopCommonModes];
+    }
+    
+    // Set default values
+    _trackingEnabled = YES;
+    _networkingEnabled = YES;
+    _crashReportingEnabled = YES;
+    _customEventTrackingEnabled = YES;
+    _enterContentZone = NO;
+    _locationTracking= YES;
+    _viewTrackingEnabled = YES;
+    _sessionTrackingEnabled = YES;
+    _loggingEnabled = NO;
+    _refreshContentZone = YES;
     
     NSURLSessionTask* task = [CountlyCommon.sharedInstance.URLSession dataTaskWithRequest:[self serverConfigRequest] completionHandler:^(NSData* data, NSURLResponse* response, NSError* error)
                               {
@@ -107,15 +335,14 @@ NSString* const kCountlySCKeySC = @"sc";
         if (error)
         {
             CLY_LOG_E(@"Error while fetching server configs: %@",error.description);
-            return;
         }
         
-        NSDictionary* serverConfigObject = serverConfigResponse[@"c"];
-        if (serverConfigObject) {
-            [self populateServerConfig:serverConfigObject];
-            [CountlyPersistency.sharedInstance storeServerConfig:serverConfigObject];
+        if (serverConfigResponse[kRConfig] != nil) {
+            [self populateServerConfig:serverConfigResponse];
+            [CountlyPersistency.sharedInstance storeServerConfig:serverConfigResponse];
         }
-    
+        
+        [self notifySdkConfigChange: config]; // if no config let stored ones to be set
     }];
     
     [task resume];
@@ -154,6 +381,93 @@ NSString* const kCountlySCKeySC = @"sc";
     }
     
     CLY_LOG_D(@"serverConfigRequest URL :%@", URL);
+}
+
+- (BOOL)trackingEnabled {
+    return _trackingEnabled;
+}
+
+- (BOOL)networkingEnabled {
+    return _networkingEnabled;
+}
+
+- (NSInteger)sessionInterval {
+    return _sessionInterval;
+}
+
+- (NSInteger)requestQueueSize {
+    return _requestQueueSize;
+}
+
+- (NSInteger)eventQueueSize {
+    return _eventQueueSize;
+}
+
+- (BOOL)crashReportingEnabled {
+    return _crashReportingEnabled;
+}
+
+- (BOOL)sessionTrackingEnabled {
+    return _sessionTrackingEnabled;
+}
+- (BOOL)loggingEnabled {
+    return _loggingEnabled;
+}
+
+- (NSInteger)limitKeyLength {
+    return _limitKeyLength;
+}
+
+- (NSInteger)limitValueSize {
+    return _limitValueSize;
+}
+
+- (NSInteger)limitSegValues {
+    return _limitSegValues;
+}
+
+- (NSInteger)limitBreadcrumb {
+    return _limitBreadcrumb;
+}
+
+- (NSInteger)limitTraceLine {
+    return _limitTraceLine;
+}
+
+- (NSInteger)limitTraceLength {
+    return _limitTraceLength;
+}
+
+- (BOOL)customEventTrackingEnabled {
+    return _customEventTrackingEnabled;
+}
+
+- (BOOL)viewTrackingEnabled {
+    return _viewTrackingEnabled;
+}
+
+- (BOOL)enterContentZone {
+    return _enterContentZone;
+}
+
+- (NSInteger)contentZoneInterval {
+    return _contentZoneInterval;
+}
+
+- (BOOL)consentRequired {
+    return _consentRequired;
+}
+
+- (NSInteger)dropOldRequestTime {
+    return _dropOldRequestTime;
+}
+
+- (BOOL)locationTrackingEnabled {
+    return _locationTracking;
+}
+
+- (BOOL)refreshContentZoneEnabled {
+    return _refreshContentZone;
 }
 
 @end